python/requirements.txt
numpy==1.23.1
scipy==1.8.1
pillow==9.4.0